backend/requirements.txt
asgiref==3.2.7
Django==3.0.4
djangorestframework==3.11.0
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.1
django-unused-media==0.1.13
